// src/interfaces/IBrand.ts
export interface IBrandObject {
    quoteDisplayInsuranceDisclaimer?: string;
    quoteDisplayDBADisclaimer?: string;
    quoteDisplayQuoteDisclaimer?: string;
    footerLogoUrl?: string;
    primaryHexColor?: string;
    displayQuotesDisclaimer?: boolean;
    cobrandingEnabled: boolean;
    brandColor: string;
    name: string;
    customPhoneNumber: string;
    faviconUrl: string;
    svgLogoUrl: string;
    brandPath: string;
    color?: any;
    reversedLogo?: boolean;
  }
  